hw3/requirements.txt
gym[atari,box2d,classic_control]==0.15.4
tensorflow-gpu==2.0.0
tensorflow-probability==0.8.0
tensorboardX==1.9
matplotlib==3.1.2
ipython==7.10.2
moviepy==1.0.0
pybullet==2.6.1
